refactor: relocate Sidenav to common components and update import paths
