+++ src/common/components/SidenavItem/SidenavItem.scss
